+++ buffer/src/main/java/io/netty/buffer/UnpooledDuplicatedByteBuf.java
@@ -16,13 +16,12 @@
 package io.netty.buffer;
 
 /**
- * A special {@link SlicedByteBuf} that can make optimizations because it knows the sliced buffer is of type
- * {@link AbstractByteBuf}.
+ * {@link DuplicatedByteBuf} implementation that can do optimizations because it knows the duplicated buffer
+ * is of type {@link AbstractByteBuf}.
  */
-final class SlicedAbstractByteBuf extends SlicedByteBuf {
-
-    SlicedAbstractByteBuf(AbstractByteBuf buffer, int index, int length) {
-        super(buffer, index, length);
+final class UnpooledDuplicatedByteBuf extends DuplicatedByteBuf {
+    UnpooledDuplicatedByteBuf(AbstractByteBuf buffer) {
+        super(buffer);
     }
 
     @Override
@@ -32,91 +31,91 @@ final class SlicedAbstractByteBuf extends SlicedByteBuf {
 
     @Override
     protected byte _getByte(int index) {
-        return unwrap()._getByte(idx(index));
+        return unwrap()._getByte(index);
     }
 
     @Override
     protected short _getShort(int index) {
-        return unwrap()._getShort(idx(index));
+        return unwrap()._getShort(index);
     }
 
     @Override
     protected short _getShortLE(int index) {
-        return unwrap()._getShortLE(idx(index));
+        return unwrap()._getShortLE(index);
     }
 
     @Override
     protected int _getUnsignedMedium(int index) {
-        return unwrap()._getUnsignedMedium(idx(index));
+        return unwrap()._getUnsignedMedium(index);
     }
 
     @Override
     protected int _getUnsignedMediumLE(int index) {
-        return unwrap()._getUnsignedMediumLE(idx(index));
+        return unwrap()._getUnsignedMediumLE(index);
     }
 
     @Override
     protected int _getInt(int index) {
-        return unwrap()._getInt(idx(index));
+        return unwrap()._getInt(index);
     }
 
     @Override
     protected int _getIntLE(int index) {
-        return unwrap()._getIntLE(idx(index));
+        return unwrap()._getIntLE(index);
     }
 
     @Override
     protected long _getLong(int index) {
-        return unwrap()._getLong(idx(index));
+        return unwrap()._getLong(index);
     }
 
     @Override
     protected long _getLongLE(int index) {
-        return unwrap()._getLongLE(idx(index));
+        return unwrap()._getLongLE(index);
     }
 
     @Override
     protected void _setByte(int index, int value) {
-        unwrap()._setByte(idx(index), value);
+        unwrap()._setByte(index, value);
     }
 
     @Override
     protected void _setShort(int index, int value) {
-        unwrap()._setShort(idx(index), value);
+        unwrap()._setShort(index, value);
     }
 
     @Override
     protected void _setShortLE(int index, int value) {
-        unwrap()._setShortLE(idx(index), value);
+        unwrap()._setShortLE(index, value);
     }
 
     @Override
     protected void _setMedium(int index, int value) {
-        unwrap()._setMedium(idx(index), value);
+        unwrap()._setMedium(index, value);
     }
 
     @Override
     protected void _setMediumLE(int index, int value) {
-        unwrap()._setMediumLE(idx(index), value);
+        unwrap()._setMediumLE(index, value);
     }
 
     @Override
     protected void _setInt(int index, int value) {
-        unwrap()._setInt(idx(index), value);
+        unwrap()._setInt(index, value);
     }
 
     @Override
     protected void _setIntLE(int index, int value) {
-        unwrap()._setIntLE(idx(index), value);
+        unwrap()._setIntLE(index, value);
     }
 
     @Override
     protected void _setLong(int index, long value) {
-        unwrap()._setLong(idx(index), value);
+        unwrap()._setLong(index, value);
     }
 
     @Override
     protected void _setLongLE(int index, long value) {
-        unwrap()._setLongLE(idx(index), value);
+        unwrap()._setLongLE(index, value);
     }
 }
