+++ buffer/src/main/java/io/netty/buffer/UnpooledSlicedByteBuf.java
@@ -16,12 +16,18 @@
 package io.netty.buffer;
 
 /**
- * {@link DuplicatedByteBuf} implementation that can do optimizations because it knows the duplicated buffer
- * is of type {@link AbstractByteBuf}.
+ * A special {@link AbstractUnpooledSlicedByteBuf} that can make optimizations because it knows the sliced buffer is of
+ * type {@link AbstractByteBuf}.
  */
-final class DuplicatedAbstractByteBuf extends DuplicatedByteBuf {
-    DuplicatedAbstractByteBuf(AbstractByteBuf buffer) {
-        super(buffer);
+final class UnpooledSlicedByteBuf extends AbstractUnpooledSlicedByteBuf {
+
+    UnpooledSlicedByteBuf(AbstractByteBuf buffer, int index, int length) {
+        super(buffer, index, length);
+    }
+
+    @Override
+    public int capacity() {
+        return maxCapacity();
     }
 
     @Override
@@ -31,91 +37,91 @@ final class DuplicatedAbstractByteBuf extends DuplicatedByteBuf {
 
     @Override
     protected byte _getByte(int index) {
-        return unwrap()._getByte(index);
+        return unwrap()._getByte(idx(index));
     }
 
     @Override
     protected short _getShort(int index) {
-        return unwrap()._getShort(index);
+        return unwrap()._getShort(idx(index));
     }
 
     @Override
     protected short _getShortLE(int index) {
-        return unwrap()._getShortLE(index);
+        return unwrap()._getShortLE(idx(index));
     }
 
     @Override
     protected int _getUnsignedMedium(int index) {
-        return unwrap()._getUnsignedMedium(index);
+        return unwrap()._getUnsignedMedium(idx(index));
     }
 
     @Override
     protected int _getUnsignedMediumLE(int index) {
-        return unwrap()._getUnsignedMediumLE(index);
+        return unwrap()._getUnsignedMediumLE(idx(index));
     }
 
     @Override
     protected int _getInt(int index) {
-        return unwrap()._getInt(index);
+        return unwrap()._getInt(idx(index));
     }
 
     @Override
     protected int _getIntLE(int index) {
-        return unwrap()._getIntLE(index);
+        return unwrap()._getIntLE(idx(index));
     }
 
     @Override
     protected long _getLong(int index) {
-        return unwrap()._getLong(index);
+        return unwrap()._getLong(idx(index));
     }
 
     @Override
     protected long _getLongLE(int index) {
-        return unwrap()._getLongLE(index);
+        return unwrap()._getLongLE(idx(index));
     }
 
     @Override
     protected void _setByte(int index, int value) {
-        unwrap()._setByte(index, value);
+        unwrap()._setByte(idx(index), value);
     }
 
     @Override
     protected void _setShort(int index, int value) {
-        unwrap()._setShort(index, value);
+        unwrap()._setShort(idx(index), value);
     }
 
     @Override
     protected void _setShortLE(int index, int value) {
-        unwrap()._setShortLE(index, value);
+        unwrap()._setShortLE(idx(index), value);
     }
 
     @Override
     protected void _setMedium(int index, int value) {
-        unwrap()._setMedium(index, value);
+        unwrap()._setMedium(idx(index), value);
     }
 
     @Override
     protected void _setMediumLE(int index, int value) {
-        unwrap()._setMediumLE(index, value);
+        unwrap()._setMediumLE(idx(index), value);
     }
 
     @Override
     protected void _setInt(int index, int value) {
-        unwrap()._setInt(index, value);
+        unwrap()._setInt(idx(index), value);
     }
 
     @Override
     protected void _setIntLE(int index, int value) {
-        unwrap()._setIntLE(index, value);
+        unwrap()._setIntLE(idx(index), value);
     }
 
     @Override
     protected void _setLong(int index, long value) {
-        unwrap()._setLong(index, value);
+        unwrap()._setLong(idx(index), value);
     }
 
     @Override
     protected void _setLongLE(int index, long value) {
-        unwrap()._setLongLE(index, value);
+        unwrap()._setLongLE(idx(index), value);
     }
 }
